User account lock out after MAX_FAILED_LOGIN_ATTEMPTS
Changes added: Support added to allow users to be locked out after a specified number of failed login attempts. This is controlled by the MAX_FAILED_LOGIN_ATTEMPTS setting in the config file. If it is set to a positive number, then the user will be locked out after that number of failed login attempts. The user will be able to log in again after the LOCKOUT_DURATION has passed. The LOCKOUT_DURATION is also set in the config file. The default is 600 seconds.

Problem: User was getting locked after reaching the MAX_FAILED_LOGIN_ATTEMPTS, But actually user needs to be locked after MAX_FAILED_LOGIN_ATTEMPTS.

Solution: Added support to allow users to be locked out after a specified number of failed login attempts. This is controlled by the MAX_FAILED_LOGIN_ATTEMPTS setting.
The details of the testing is given below MR
https://gitlab-master.nvidia.com/dgx/bmc/bmcweb/-/merge_requests/760

Fixes nvbug https://nvbugs/4001568