user-mgr: improved error handling
adds some error handling to createUser, deleteUser and renameUser.

createUser: if useradd fails but the user exists, try to delete it.
deleteUser: if userdel fails and the user does not exist, treat as
success. Also add `-f` to userdel.
renameUser: if the rename operation fails but results in the new user
existing, track the rename as having occurred.

Testing: added unit tests, tested against the romulus image.

Change-Id: I6ad589ca03f1dad2da21560ad1c653795d2d6427
Signed-off-by: Joey Berkovitz <joey@berkovitz.us>